requirements-dev.txt
pandas==1.0.5
pyaml
ipykernel
keplergl
pyathena
fire
networkx<=2.0.0
rdp
jinja2
jupyter
jupyter_contrib_nbextensions
pymsteams
loguru
gspread
oauth2client
df2gspread
osm_road_length
awswrangler
timezonefinder
babelgrid
matplotlib
plotly
cufflinks
seaborn
plotnine
siuba
adtk
pyod
tsmoothie
osmpy
geopandas
openpyxl
folium
contextily